Split FP position sums in wclevelcont_func10 for operand order
dll_028D_wclevelcont: precompute (lbl_803E6DD0 + px) into a temp before
adding the third term so the inner fadds emits before the trailing
constant load, matching target across all six block-origin arms.
82.51% -> 86.14%.

+++ src/main/dll/duster.c
@@ -891,7 +891,7 @@ void fn_80156950(uint obj, int state)
 
 void mutatedEbaUpdateWhileFrozen(uint obj, int state, undefined4 unused, int eventKind)
 {
-    short move;
+    int move;
 
     if (eventKind != 0x11)
     {
